js refactor
