+++ crates/binlex_tools/binlex-web/src/assets.rs
@@ -18,15 +18,31 @@ pub(crate) const SCRIPT: &str = concat!(
     "\n",
     include_str!("assets/js/features/query_assistant.js"),
     "\n",
-    include_str!("assets/js/features/results_controls.js"),
+    include_str!("assets/js/features/results_row_actions.js"),
     "\n",
-    include_str!("assets/js/features/results_popovers.js"),
+    include_str!("assets/js/features/results_columns.js"),
     "\n",
-    include_str!("assets/js/features/upload.js"),
+    include_str!("assets/js/features/results_popovers_shared.js"),
     "\n",
-    include_str!("assets/js/features/auth_profile.js"),
+    include_str!("assets/js/features/results_popovers_corpora.js"),
     "\n",
-    include_str!("assets/js/features/admin.js"),
+    include_str!("assets/js/features/results_popovers_tags_symbols.js"),
+    "\n",
+    include_str!("assets/js/features/results_popovers_comments.js"),
+    "\n",
+    include_str!("assets/js/features/upload_modal.js"),
+    "\n",
+    include_str!("assets/js/features/upload_status.js"),
+    "\n",
+    include_str!("assets/js/features/auth.js"),
+    "\n",
+    include_str!("assets/js/features/two_factor.js"),
+    "\n",
+    include_str!("assets/js/features/profile.js"),
+    "\n",
+    include_str!("assets/js/features/admin_users.js"),
+    "\n",
+    include_str!("assets/js/features/admin_metadata.js"),
     "\n",
     include_str!("assets/app.js"),
 );
